Visual Clarity Update Changed the default color of grid to red. This makes "unsolved" blocks very obvious and easy to find. When solving each block, the color will change to black, this includes empty blocks, numbered blocks, and flags.
Commonly near the end of the game, you are frantically searching for the last few blocks to solve in time and it is very hard to quickly find the ? tiles when it blends in with all the other blocks.

This change should improve the visual clarity of those unsolved blocks and make it much easier to find. The challenge of the game should be logic, decision making, and risk; not being able to scan through bad visuals.